# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Class Pets API
  version: 1.0.0
  description: 班级宠物积分系统接口文档（Spring Boot + MyBatis-Plus + MySQL）。注意：本文档含规划接口与已实现接口，联调请以 PROJECT_STATUS.md 中“当前接口（真实）”为准。
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "200":
          description: ok
  /auth/login:
    post:
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: ok
...
  /classes:
    get:
      tags: [Class]
      responses:
        "200": { description: ok }
    post:
      tags: [Class]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassCreateRequest'
      responses:
        "200": { description: ok }
  /classes/{classId}:
    get:
      tags: [Class]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      responses:
        "200": { description: ok }
    put:
      tags: [Class]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassUpdateRequest'
      responses:
        "200": { description: ok }
    delete:
      tags: [Class]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      responses:
        "200": { description: ok }
...
  /classes/{classId}/data/export/students:
    get:
      tags: [Data]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      responses:
        "200": { description: ok }
  /classes/{classId}/data/export/rules:
    get:
      tags: [Data]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      responses:
        "200": { description: ok }
  /classes/{classId}/data/export/events:
    get:
      tags: [Data]
      parameters:
        - { $ref: '#/components/parameters/ClassId' }
        - in: query
          name: from
          schema: { type: integer, format: int64 }
          example: 1738368000000
        - in: query
          name: to
          schema: { type: integer, format: int64 }
          example: 1738972799000
      responses:
        "200": { description: ok }
  /classes/{classId}/data/import/students/preview:
    post:
      tags: [Data]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: ok
          content:
            application/json:
              example:
                code: 0
                message: "OK"
                data:
                  totalRows: 3
                  validRows: 2
                  errorRows: 1
                  rows:
                    - lineNo: 2
                      name: "张小明"
                      studentNo: "202601"
                      gender: "男"
                      groupName: "火箭队"
                      points: "10"
                      error: ""
                    - lineNo: 3
                      name: ""
                      studentNo: "202602"
                      gender: "女"
                      groupName: "火箭队"
                      points: "8"
                      error: "姓名不能为空"
  /classes/{classId}/data/import/students/commit:
    post:
      tags: [Data]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentImportCommitRequest'
            example:
              rows:
                - lineNo: 2
                  name: "张小明"
                  studentNo: "202601"
                  gender: "男"
                  groupName: "火箭队"
                  points: "10"
                  error: ""
      responses:
        "200":
          description: ok
          content:
            application/json:
              example:
                code: 0
                message: "OK"
                data:
                  total: 3
                  created: 2
                  updated: 0
                  failed: 1
                  errors:
                    - lineNo: 3
                      studentNo: "202602"
                      message: "姓名不能为空"
  /classes/{classId}/data/import/students/errors-export:
    post:
      tags: [Data]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentImportCommitRequest'
      responses:
        "200": { description: ok }
  /classes/{classId}/data/import/students/template:
    get:
      tags: [Data]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      responses:
        "200": { description: ok }
  /classes/{classId}/data/import/rules/preview:
    post:
      tags: [Data]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: ok
          content:
            application/json:
              example:
                code: 0
                message: "OK"
                data:
                  totalRows: 3
                  validRows: 2
                  errorRows: 1
                  rows:
                    - lineNo: 2
                      content: "课堂积极发言"
                      points: "2"
                      type: "加分"
                      targetType: "个人"
                      category: "课堂表现"
                      cooldownHours: "0"
                      stackable: "是"
                      error: ""
                    - lineNo: 3
                      content: ""
                      points: "3"
                      type: "扣分"
                      targetType: "小组"
                      category: "作业管理"
                      cooldownHours: "24"
                      stackable: "否"
                      error: "规则内容不能为空"
  /classes/{classId}/data/import/rules/commit:
    post:
      tags: [Data]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleImportCommitRequest'
            example:
              conflictPolicy: overwrite
              rows:
                - lineNo: 2
                  content: "课堂积极发言"
                  points: "2"
                  type: "加分"
                  targetType: "个人"
                  category: "课堂表现"
                  cooldownHours: "0"
                  stackable: "是"
                  error: ""
      responses:
        "200":
          description: ok
          content:
            application/json:
              example:
                code: 0
                message: "OK"
                data:
                  total: 4
                  created: 2
                  updated: 1
                  skipped: 0
                  failed: 1
                  errors:
                    - lineNo: 5
                      content: "作业迟交"
                      message: "分值必须大于0"
  /classes/{classId}/data/import/rules/errors-export:
    post:
      tags: [Data]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleImportCommitRequest'
      responses:
        "200": { description: ok }
  /classes/{classId}/data/import/rules/template:
    get:
      tags: [Data]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      responses:
        "200": { description: ok }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ClassId: { in: path, name: classId, required: true, schema: { type: integer, format: int64 } }
...
  schemas:
    RegisterRequest:
      type: object
      required: [username, password, activationCode]
      properties:
        username: { type: string }
        password: { type: string }
        nickname: { type: string }
        activationCode: { type: string }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
...
    ClassCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        teacherName: { type: string }
    ClassUpdateRequest:
      type: object
      properties:
        name: { type: string }
        teacherName: { type: string }
...
    StudentImportCommitRequest:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: '#/components/schemas/StudentImportRow'
      example:
        rows:
          - lineNo: 2
            name: "张小明"
            studentNo: "202601"
            gender: "男"
            groupName: "火箭队"
            points: "10"
            error: ""
          - lineNo: 3
            name: ""
            studentNo: "202602"
            gender: "女"
            groupName: "火箭队"
            points: "8"
            error: "姓名不能为空"
    StudentImportRow:
      type: object
      properties:
        lineNo: { type: integer }
        name: { type: string }
        studentNo: { type: string }
        gender: { type: string }
        groupName: { type: string }
        points: { type: string }
        error: { type: string }
    RuleImportCommitRequest:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: '#/components/schemas/RuleImportRow'
        conflictPolicy:
          type: string
          enum: [skip, overwrite]
          default: skip
      example:
        conflictPolicy: skip
        rows:
          - lineNo: 2
            content: "课堂积极发言"
            points: "2"
            type: "加分"
            targetType: "个人"
            category: "课堂表现"
            cooldownHours: "0"
            stackable: "是"
            error: ""
          - lineNo: 3
            content: ""
            points: "3"
            type: "扣分"
            targetType: "小组"
            category: "作业管理"
            cooldownHours: "24"
            stackable: "否"
            error: "规则内容不能为空"
    RuleImportRow:
      type: object
      properties:
        lineNo: { type: integer }
        content: { type: string }
        points: { type: string }
        type: { type: string }
        targetType: { type: string }
        category: { type: string }
        cooldownHours: { type: string }
        stackable: { type: string }
        error: { type: string }
...